DPAD-995 :: Moved around the play state components
